// app/actions/browser/index.ts
import {
  BrowserActionType,
  AddToViewedDappAction,
  AddToBrowserHistoryAction,
  AddToBrowserWhitelistAction,
  ClearBrowserHistoryAction,
  CloseAllTabsAction,
  CreateNewTabAction,
  CloseTabAction,
  SetActiveTabAction,
  UpdateTabAction,
  StoreFaviconUrlAction,
} from './types';

export * from './types';

/**
 * Adds a new entry to viewed dapps
 *
 * @param hostname - Dapp hostname
 */
export function addToViewedDapp(hostname: string): AddToViewedDappAction {
  return {
    type: BrowserActionType.ADD_TO_VIEWED_DAPP,
    hostname,
  };
}

/**
 * Adds a new entry to the browser history
 *
 * @param website - The website that has been visited
 * @param website.url - The website's url
 * @param website.name - The website name
 */
export function addToHistory({ url, name }: { url: string; name: string }): AddToBrowserHistoryAction {
  return {
    type: BrowserActionType.ADD_TO_BROWSER_HISTORY,
    url,
    name,
  };
}

/**
 * Clears the entire browser history
 */
export function clearHistory(metricsEnabled: boolean, marketingEnabled: boolean): ClearBrowserHistoryAction {
  return {
    type: BrowserActionType.CLEAR_BROWSER_HISTORY,
    id: Date.now(),
    metricsEnabled,
    marketingEnabled,
  };
}

/**
 * Adds a new entry to the whitelist
 *
 * @param url - The website's url
 */
export function addToWhitelist(url: string): AddToBrowserWhitelistAction {
  return {
    type: BrowserActionType.ADD_TO_BROWSER_WHITELIST,
    url,
  };
}

/**
 * Closes all the opened tabs
 */
export function closeAllTabs(): CloseAllTabsAction {
  return {
    type: BrowserActionType.CLOSE_ALL_TABS,
  };
}

/**
 * Creates a new tab
 *
 * @param url - The website's url
 * @param linkType - optional link type
 */
export function createNewTab(url: string, linkType?: string): CreateNewTabAction {
  return {
    type: BrowserActionType.CREATE_NEW_TAB,
    url,
    linkType,
    id: Date.now(),
  };
}

/**
 * Closes an exiting tab
 *
 * @param id - The Tab ID
 */
export function closeTab(id: number): CloseTabAction {
  return {
    type: BrowserActionType.CLOSE_TAB,
    id,
  };
}

/**
 * Selects an exiting tab
 *
 * @param id - The Tab ID
 */
export function setActiveTab(id: number): SetActiveTabAction {
  return {
    type: BrowserActionType.SET_ACTIVE_TAB,
    id,
  };
}

/**
 * Updates an existing tab
 *
 * @param id - The Tab ID
 * @param data - Tab data to update
 */
export function updateTab(id: number, data: { isArchived?: boolean; url?: string; image?: string }): UpdateTabAction {
  return {
    type: BrowserActionType.UPDATE_TAB,
    id,
    data,
  };
}

/**
 * Stores the favicon url using the origin as key
 * @param favicon - favicon to store
 * @param favicon.origin - the origin of the favicon as key
 * @param favicon.url - the favicon image url
 */
export function storeFavicon({ origin, url }: { origin: string; url: string }): StoreFaviconUrlAction {
  return {
    type: BrowserActionType.STORE_FAVICON_URL,
    origin,
    url,
  };
}
